requirements.txt
absl-py==2.0.0
accelerate==0.29.1
addict==2.4.0
aiohttp==3.9.1
aioprometheus==23.3.0
aiosignal==1.3.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
base58==2.1.1
beartype==0.14.1
beautifulsoup4==4.12.2
bitsandbytes==0.41.3.post2
black==24.3.0
bleach==6.1.0
blingfire==0.1.8
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorcet==3.1.0
comm==0.2.0
contourpy==1.2.0
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
datasets==2.15.0
debugpy==1.8.0
decorator==5.1.1
deepspeed==0.12.4
defusedxml==0.7.1
dill==0.3.7
diskcache==5.6.3
distro==1.8.0
docker-pycreds==0.4.0
docstring-parser==0.15
einops==0.7.0
evaluate==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
fancy-einsum==0.0.3
fastapi==0.106.0
fastjsonschema==2.19.0
filelock==3.13.1
flash-attn==2.3.6
fonttools==4.46.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.26.0
google-auth-oauthlib==1.2.0
grpcio==1.60.0
h11==0.14.0
higher==0.2.1
hjson==3.1.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.27.1
hydra-core==1.3.2
idna==3.6
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
jaxtyping==0.2.24
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json5==0.9.14
jsonlines==4.0.0
jsonnet==0.20.0
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
loguru==0.7.2
lxml==5.1.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
modelscope==1.19.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
ninja==1.11.1.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.2
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.3.7
orjson==3.9.10
oss2==2.19.0
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
peft==0.7.0
pexpect==4.9.0
Pillow==9.3.0
platformdirs==4.1.0
plotly==5.22.0
portalocker==2.8.2
POT==0.9.1
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.21.0
pydantic==1.10.13
pydantic_core==2.14.5
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
quantile-python==1.1
rapidfuzz==3.6.2
ray==2.9.0
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rouge==1.0.1
rouge-score==0.1.2
rpds-py==0.13.2
rsa==4.9
ruff==0.3.3
sacrebleu==2.4.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
Send2Trash==1.8.2
sentence-transformers==2.7.0
sentencepiece==0.1.99
sentry-sdk==1.38.0
setproctitle==1.3.3
shtab==1.6.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.2.0
tiktoken==0.6.0
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
#torch==2.1.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformer-lens==1.11.0
#transformers==4.43.0
triton==2.1.0
trl==0.8.1
typeguard==2.13.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.8.0
tyro==0.6.0
tzdata==2023.3
ujson==5.9.0
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
vllm==0.2.6
wandb==0.16.1
watchfiles==0.21.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wget==3.2
widgetsnbextension==4.0.9
xformers==0.0.23.post1
xxhash==3.4.1
yarl==1.9.3
